feat: Created separate file for homepage, created error display component and UI changes in login page
